Extract an abstract class 'CardinalityBasedJoinPlanOptimizerBase'
